[Fix] refactor structure
