+++ src/assets/styles/variables.scss
@@ -1,0 +1,5 @@
+$color_main_text: #1d1d1b;
+$color_second_text: #6d727c;
+
+$primary-color: #282c34;
+$secondary-color: #61dafb;
