+++ src/index.scss
@@ -1,3 +1,5 @@
+@import "./assets/styles/variables.scss";
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
@@ -19,5 +21,3 @@ code {
   text-decoration: none;
   color: inherit;
 }
-
-.DayPicker-Day {}
