--- src/styles/theme.ts
@@ -1,2 +1,0 @@
-export const theme = {}
-export const ThemeProvider = {}
