--- src/view/common/Button/styles.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-type PropsType = {
-  type?: string;
-};
-
-export const Wrap = styled.div<PropsType>`
-  width: 100%;
-`;
