--- src/view/common/Button/types.ts
@@ -1,3 +1,0 @@
-export type PropsType = {
-    type?: string;
-};
