+++ src/view/pages/MainPage/index.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import MultiDatePicker from "../../common/MultiDatePicker";
 import GoogleLogin from '../../components/GoogleLogin';
 import FacebookLogin from '../../components/FacebookLogin';
+import Button from "../../common/Button/index";
 
 const  MainPage = () => {
   return (
@@ -11,6 +12,7 @@ const  MainPage = () => {
       <MultiDatePicker/>
       <GoogleLogin/>
       <FacebookLogin/>
+      <Button variant="outlined">Button</Button>
     </div>
   );
 }
